+++ integration-tests/upgrade/upgrade.go
@@ -53,10 +53,10 @@ func TestUpgrade(ctx context.Context, t testing.T, chain testing.Chain) {
 		},
 	))
 	requireT.NoError(err)
-	log.Info("Upgrade proposal has been submitted", zap.Int("proposalID", proposalID))
+	log.Info("Upgrade proposal has been submitted", zap.Uint64("proposalID", proposalID))
 
 	// Verify that voting period started.
-	proposal, err := chain.Governance.GetProposal(ctx, uint64(proposalID))
+	proposal, err := chain.Governance.GetProposal(ctx, proposalID)
 	requireT.NoError(err)
 	requireT.Equal(govtypes.StatusVotingPeriod, proposal.Status)
 
@@ -67,7 +67,7 @@ func TestUpgrade(ctx context.Context, t testing.T, chain testing.Chain) {
 	log.Info("Voters have voted successfully, waiting for voting period to be finished", zap.Time("votingEndTime", proposal.VotingEndTime))
 
 	// Wait for proposal result.
-	finalStatus, err := chain.Governance.WaitForVotingToFinalize(ctx, uint64(proposalID))
+	finalStatus, err := chain.Governance.WaitForVotingToFinalize(ctx, proposalID)
 	requireT.NoError(err)
 	requireT.Equal(govtypes.StatusPassed, finalStatus)
